python_base_04/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Limiter==2.7.0
Flask-CORS==3.0.10
Flask-SocketIO==5.3.6
Jinja2==3.1.2
PyJWT==2.6.0
bcrypt==4.3.0
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pymongo==4.6.1
redis==5.2.1
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.2
cryptography>=41.0.0
APScheduler==3.10.4
PyYAML==6.0.1
stripe==7.11.0
